Type consistency fixes and Cortex-M0+ assembly optimizations
- Fix type inconsistencies across the codebase (uint8_t, size_t, uint32_t)
- Add const qualifiers to read-only variables and parameters
- Add optimized assembly math functions for Cortex-M0+ (smul64, usqr64)
- Add utilAtoui for unsigned string-to-integer conversion
- Fix printf format specifiers to match argument types
- Use size_t for array index loop counters
- Place asm math functions in RAM for faster execution

+++ src/driver_DMAC.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stddef.h>
+
 #include "emon32_samd.h"
 
 typedef struct DMACCfgCh {
@@ -13,7 +15,7 @@ void dmacSetup(void);
  *  @param [in] ch : channel
  *  @return Pointer to the DmacDescriptor struct
  */
-volatile DmacDescriptor *dmacGetDescriptor(uint32_t ch);
+volatile DmacDescriptor *dmacGetDescriptor(uint8_t ch);
 
 /*! @brief Set the callback when the DMA has filled the sample buffer
  *  @param [in] cb : pointer to the callback function
@@ -28,37 +30,37 @@ void dmacCallbackUartCmpl(void (*cb)(void));
 /*! @brief Disable a DMAC channel
  *  @param [in] ch : channel number
  */
-void dmacChannelDisable(uint32_t ch);
+void dmacChannelDisable(uint8_t ch);
 
 /*! @brief Enable a DMAC channel
  *  @param [in] ch : channel number
  */
-void dmacChannelEnable(uint32_t ch);
+void dmacChannelEnable(uint8_t ch);
 
 /*! @brief Configure a DMA channel
  *  @param [in] ch : channel to configure
  *  @param [in] pCfg : pointer to configuration details
  */
-void dmacChannelConfigure(uint32_t ch, const DMACCfgCh_t *pCfg);
+void dmacChannelConfigure(uint8_t ch, const DMACCfgCh_t *pCfg);
 
 /*! @brief Enable DMAC channel interrupt
  *  @param [in] ch : channel to enable interrupt for
  */
-void dmacEnableChannelInterrupt(uint32_t ch);
+void dmacEnableChannelInterrupt(uint8_t ch);
 
 /*! @brief Disable DMAC channel interrupt
  *  @param [in] ch : channel to disable interrupt for
  */
-void dmacDisableChannelInterrupt(uint32_t ch);
+void dmacDisableChannelInterrupt(uint8_t ch);
 
 /*! @brief Clear DMAC channel interrupt flag
  *  @param [in] ch : channel to clear interrupt flag for
  */
-void dmacClearChannelInterrupt(uint32_t ch);
+void dmacClearChannelInterrupt(uint8_t ch);
 
 /*! @brief Calculate the CRC16 (CCITT - 0x1021)
  *  @param [in] pData : pointer to data
  *  @param [in] n : number of bytes in data
  *  @return CRC16 value
  */
-uint16_t calcCRC16_ccitt(const void *pSrc, uint32_t n);
+uint16_t calcCRC16_ccitt(const void *pSrc, size_t n);
